repo: refactor into package.
The first step to fixing the stage caching issue is to refactor the repo
into a package so each class can live in its own module.

LP: #1661383
