+++ snapcraft/internal/repo/manifest.txt
